requirements.txt
numpy
pandas
selenium
unidecode
webdriver-manager
PyYAML
xlwt